move make_non_pad_mask to nets_utils
